third_party/blink/web_tests/svg/dynamic-updates/SVGCircleElement-svgdom-r-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'r' property of the SVGCircleElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS circleElement.r.baseVal.value is 1
PASS circleElement.r.baseVal.value is 150
PASS successfullyParsed is true

TEST COMPLETE
